refactor: 💡 move mock-data
