+++ sagaicle/api/mocks/likes.mock.ts
@@ -1,4 +1,4 @@
-import type { IsLikedResponse, LikeResponse, DislikeResponse } from "@/api";
+import type { IsLikedResponse, LikeResponse, DislikeResponse } from "../types";
 
 const isLikedResponse: IsLikedResponse = {
   route_id: "12345678-1234-1234-1234-1234567890ab",
